Modified node storage to use fixed precision instead of doubles to reduce storage space. Represented some ints as chars for further space reduction.

+++ src/com/bretth/osmosis/core/bdb/v0_5/impl/DatabaseRelationIterator.java
@@ -76,7 +76,9 @@ public class DatabaseRelationIterator implements ReleasableIterator<Long> {
 			
 			// We want to start retrieving records from the first record with a
 			// part 1 of searchId. Therefore part2 must be the minimum value.
-			startKey = new LongLongIndexElement(searchId, Long.MIN_VALUE);
+			// HACK: Note that this should be using Long.MIN_VALUE but the
+			// internal implementation is currently using an integer.
+			startKey = new LongLongIndexElement(searchId, Integer.MIN_VALUE);
 			keyBinding.objectToEntry(startKey, keyEntry);
 			
 			cursor = db.openCursor(txn, null);
